requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
colorama==0.4.6
diff-match-patch==20241021
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.7
django-simple-history==3.8.0
djangorestframework==3.16.0
Faker==37.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
pycryptodome==3.22.0
sqlparse==0.5.3
tablib==3.8.0
toposort==1.10
tzdata==2025.2
whitenoise==6.9.0
